Extract parser backup and restore grammar
